Merge all arb issuance data, as opposed to subset
There is a small subset of credits data that we only source from the ARB
issuance table. This means we want to patch in all credit data from the
issuance table, as opposed to only including credits data from projects
for which we've also source crediting data from one of the registries.